Removed test button on main page. Added functions and logic to the Debtor Device - Send Button
