+++ components/Device/Debtor.tsx
@@ -1,7 +1,40 @@
-import { TimeComponent } from "components/timeComponent/TimeComponent"
+import axios from "axios"
 import Image from "next/image"
+import { useContext } from "react"
+import { TimeComponent } from "components/timeComponent/TimeComponent"
+import EntityContext from "store/entities/entity.context"
 
 export function DebtorDevice() {
+  const entityCtx = useContext(EntityContext)
+  const postPacs002Test = async () => {
+    try {
+      const response = await axios.post(
+        "http://localhost:5001/v1/evaluate/iso20022/pacs.002.001.12",
+        entityCtx.pacs002,
+        { headers: { "Content-Type": "application/json" } }
+      )
+      console.log("Test POST PACS002 response: ", response.data)
+    } catch (error) {
+      console.log(error)
+    }
+  }
+
+  const postPacs008Test = async () => {
+    try {
+      const response = await axios.post(
+        "http://localhost:5001/v1/evaluate/iso20022/pacs.008.001.10",
+        entityCtx.pacs008,
+        { headers: { "Content-Type": "application/json" } }
+      )
+
+      if (response.status === 200) {
+        await postPacs002Test()
+      }
+      console.log("Test POST PACS008 response: ", response.data)
+    } catch (error) {
+      console.log(error)
+    }
+  }
   return (
     <div className="relative col-span-4" style={{ height: "505px" }}>
       <Image src="/device.svg" width="250" height="505" className="absolute left-20 top-0 " alt="" priority={true} />
@@ -47,8 +80,13 @@ export function DebtorDevice() {
       </div>
 
       <div className="absolute" style={{ marginLeft: "94px", width: "222px", bottom: "25px" }}>
-        <div className="bg-black ml-5 rounded-lg text-white w-4/5" style={{ padding: ".1em" }}>
-          <button className="border border-white p-1 rounded-lg w-full">
+        <div className="ml-5 w-4/5 rounded-lg bg-black text-white" style={{ padding: ".1em" }}>
+          <button
+            className="w-full rounded-lg border border-white p-1"
+            onClick={async () => {
+              await postPacs008Test()
+            }}
+          >
             Send
           </button>
         </div>
